CO-191 refactor spark policy fixtures
