+++ docs/PRD-linear-aaa8be94-c82b-40a8-bb19-48343183790e.md
@@ -1,0 +1,129 @@
+# PRD - CO: Refactor docs-hygiene spark policy guard into parser-backed fixture harness
+
+- Linear issue: `CO-191` / `aaa8be94-c82b-40a8-bb19-48343183790e`
+- Source issue: `CO-183` / `df69fabe-63c2-4b98-a226-9c37892b4f9d`
+- Problem Statement: CO-183 preserved the immediate `explorer_fast` / `gpt-5.3-codex-spark` file/codebase-search-only policy with targeted regressions, but the docs-hygiene guard is now difficult to audit because policy wording is encoded in clause/window heuristics and inline test phrases.
+- Desired Outcome: preserve CO-183's accepted production behavior while moving spark-policy classification into a focused parser-backed classifier and a structured fixture harness with stable ids, expected verdicts, and failure reasons.
+
+## User Request Translation (Context Anchor)
+- User intent / needs: replace the regex-heavy spark-policy model-posture check with a small deterministic classifier and fixture-driven evaluator so future wording changes are reviewed by editing structured cases instead of growing ad hoc heuristics.
+- Success criteria / acceptance:
+  - existing CO-183 behavior is preserved for the current docs and existing tests
+  - classifier concepts are explicit: marker, clause, assertion window, scope qualifier, forbidden usage, restrictive non-use, and neutral reference
+  - spark-policy phrase coverage lives in structured fixtures with stable ids and expected reasons
+  - `npm run docs:check` still uses the same production rule as the fixture harness
+  - focused docs-hygiene tests, docs:check, and scoped guardrails pass
+- Constraints / non-goals:
+  - do not change the spark policy itself
+  - do not weaken file/codebase-search-only restrictions for `explorer_fast`
+  - do not reopen Codex CLI 0.120 adoption posture
+  - do not touch docs:freshness or CO-184/CO-188 freshness maintenance
+  - do not introduce a broad natural-language parser for all docs hygiene rules
+
+## Intent Checksum
+- Exact user wording / phrases to preserve:
+  - `preserve CO-183's finalized spark file/codebase-search-only policy behavior`
+  - `parser-backed classifier`
+  - `structured fixture harness`
+  - `future policy wording changes are audited through structured fixtures`
+- Protected terms / exact artifact and surface names:
+  - `scripts/docs-hygiene.ts`
+  - `scripts/lib/docs-catalog.js`
+  - `tests/docs-hygiene.spec.ts`
+  - `tests/fixtures/docs-hygiene/spark-policy-cases.json`
+  - `docs:check`
+  - `explorer_fast`
+  - `gpt-5.3-codex-spark`
+  - `file/codebase search only`
+- Nearby wrong interpretations to reject:
+  - reopening Codex CLI 0.120 adoption posture
+  - changing or weakening the spark policy
+  - blocking CO-183 after its immediate regressions are fixed
+  - building a broad natural-language parser for every docs-hygiene rule
+  - storing raw Linear workpads, private review transcripts, or unsanitized operator text as fixtures
+
+## Parity / Alignment Matrix
+- Current truth:
+  - docs-hygiene model posture checks allow the current primary `gpt-5.4` posture, the `explorer_fast` `gpt-5.3-codex-spark` exception, and known unsupported-review warning forms
+  - the spark exception rule depends on line-level detection and inline regression phrases
+  - adding new allowed, forbidden, or neutral cases requires editing test bodies and guard heuristics together
+- Reference truth:
+  - CO-183 final behavior defines the accepted file/codebase-search-only spark exception policy
+  - `docs/guides/codex-version-policy.md` remains the current posture source
+  - existing `tests/docs-hygiene.spec.ts` cases encode the current regression floor
+- Target truth / intended delta:
+  - the production model-posture rule delegates spark-policy classification to a focused parser-backed classifier
+  - fixtures encode stable ids, categories, expected verdicts, and expected reasons
+  - table-driven tests report fixture id, expected verdict, actual verdict, and reason on mismatch
+  - `docs:check` and tests share the same classifier path
+- Explicitly out-of-scope differences:
+  - no Codex CLI adoption posture changes
+  - no docs:freshness maintenance work
+  - no weakening of `explorer_fast` restrictions
+  - no semantic parser beyond this bounded spark-policy classifier
+
+## Not Done If
+- The spark policy classifier remains an expanding regex/window heuristic with the main phrase corpus embedded in a single test body.
+- Future allowed/forbidden/neutral cases cannot be added as structured fixtures with stable ids and failure reasons.
+- The refactor changes CO-183's final accepted production behavior without explicit fixture evidence and migration notes.
+- The `docs:check` production rule diverges from the fixture/evaluator path.
+- The issue broadens into unrelated docs:freshness, Codex 0.120 adoption, or general docs-language parsing.
+
+## Goals
+- Create the CO-191 docs-first packet and registry mirrors.
+- Extract a focused spark-policy classifier used by docs-hygiene production checks.
+- Move spark-policy regression phrases into structured fixtures.
+- Add table-driven fixture tests with high-signal mismatch output.
+- Validate the existing docs remain compliant through `npm run docs:check`.
+
+## Non-Goals
+- Change the active model posture or spark policy.
+- Add support for broad natural-language parsing.
+- Touch docs:freshness, CO-184, or CO-188 maintenance.
+- Use private Linear workpads or unsanitized operator text as fixtures.
+
+## Stakeholders
+- Product: CO operators relying on docs truthfulness gates.
+- Engineering: maintainers of docs-hygiene, docs-catalog helpers, and provider-worker validation.
+- Design: not applicable.
+
+## Metrics & Guardrails
+- Primary Success Metrics:
+  - fixture evaluator passes all structured spark-policy cases
+  - focused docs-hygiene tests and `npm run docs:check` pass
+  - existing model posture failures retain their current error wording where tested
+- Guardrails / Error Budgets:
+  - fail closed on stale non-spark model posture references
+  - keep `explorer_fast` allowed only for restrictive file/codebase-search-only spark references
+  - preserve unsupported-review warning handling
+
+## User Experience
+- Personas:
+  - maintainer adding future spark-policy wording
+  - reviewer checking that policy behavior did not drift
+  - provider worker running docs:check before handoff
+- User Journeys:
+  - maintainer adds a fixture with a stable id and sees a deterministic failure reason if behavior is wrong
+  - reviewer reads the fixture diff instead of a large inline phrase corpus
+  - provider worker runs the normal docs gate and exercises the same classifier used by tests
+
+## Technical Considerations
+- Architectural Notes:
+  - keep the classifier in `scripts/lib/` near docs-catalog posture parsing
+  - classify only model-posture/spark-policy lines, not arbitrary prose
+  - normalize Markdown emphasis and wrapped continuation lines before classification
+  - keep `scripts/docs-hygiene.ts` consuming docs-catalog helpers rather than creating a separate command path
+- Dependencies / Integrations:
+  - `scripts/docs-hygiene.ts`
+  - `scripts/lib/docs-catalog.js`
+  - `scripts/lib/docs-catalog.d.ts`
+  - `tests/docs-hygiene.spec.ts`
+  - `tests/fixtures/docs-hygiene/spark-policy-cases.json`
+
+## Open Questions
+- None blocking. The first implementation pass should preserve current behavior exactly and record any intentionally deferred readability or fixture taxonomy refinements as follow-ups instead of broadening scope.
+
+## Approvals
+- Product: provider-worker lane from Linear issue `CO-191`
+- Engineering: pending docs-review / implementation review
+- Design: not applicable
